expo sdk 51 업그레이드,호환성 / components 폴더  재지정
