+++ App.jsx
@@ -1,19 +1,15 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
-import styled from 'styled-components';
-import Header from './components/Header';
-import Footer from './components/Footer';
-import Container from './components/Container';
-import Button from './components/Button';
-import '/Users/soss/Desktop/cross/cross-app/css/common.css';
+import { StatusBar } from "expo-status-bar";
+import { StyleSheet, Text, View } from "react-native";
+import "./css/common.css";
+import Header from "./src/components/Header";
+import Container from "./src/components/Container";
+import Footer from "./src/components/Footer";
 
 export default function App() {
   return (
     <View style={styles.container}>
       <Header></Header>
-      <Container>
-        
-      </Container>
+      <Container></Container>
       <Footer></Footer>
     </View>
   );
@@ -22,8 +18,8 @@ export default function App() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
   },
 });
