--- components/Container.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Button from "./Button";
-const Container = () => {
-  return (
-  <div>
-    <h1>어디서 만나면 좋을까요?</h1>
-    <Button></Button>
-  </div>
-  )
-};
-
-export default Container;
