--- components/Footer.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return (
-  <div className="footer">
-푸터입니다
-  </div>
-  )
-};
-
-export default Footer;
