--- components/Header.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Header = () => {
-  return <div className="header">
-    헤더입니다
-  </div>
-};
-
-export default Header;
