+++ package.json
@@ -10,11 +10,12 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~3.2.1",
-    "expo": "~51.0.5",
+    "expo": "^51.0.8",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.1",
+    "react-native-maps": "1.14.0",
     "react-native-web": "~0.19.10",
     "styled-components": "^6.1.11"
   },
