+++ src/components/Footer.jsx
@@ -1,0 +1,5 @@
+const Footer = () => {
+  return <div className="footer">푸터입니다</div>;
+};
+
+export default Footer;
