+++ src/components/Header.jsx
@@ -1,0 +1,5 @@
+const Header = () => {
+  return <div className="header">헤더입니다</div>;
+};
+
+export default Header;
